09_FilesAndNetwork/Task_9.14/ParseJsonWithJackson/src/main/java/Station.java
import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.EqualsAndHashCode;
import lombok.ToString;

import java.io.Serializable;

@EqualsAndHashCode
@ToString
public class Station implements Serializable {
    private String name;

    public Station(@JsonProperty("name") String name) {
        this.name = name;
    }

    @JsonCreator
    public Station() {
        super();
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}